dpmi: remove some unsigned longs

+++ src/dosext/dpmi/dmemory.h
@@ -34,7 +34,7 @@ typedef struct dpmi_pm_block_root_struc {
   dpmi_pm_block *first_pm_block;
 } dpmi_pm_block_root;
 
-dpmi_pm_block *lookup_pm_block(dpmi_pm_block_root *root, unsigned long h);
+dpmi_pm_block *lookup_pm_block(dpmi_pm_block_root *root, unsigned h);
 dpmi_pm_block *lookup_pm_block_by_addr(dpmi_pm_block_root *root,
 	dosaddr_t addr);
 dpmi_pm_block *lookup_pm_block_by_shmname(dpmi_pm_block_root *root,
@@ -45,7 +45,7 @@ dpmi_pm_block *DPMI_malloc(dpmi_pm_block_root *root, unsigned int size);
 dpmi_pm_block *DPMI_mallocLinear(dpmi_pm_block_root *root, unsigned int base, unsigned int size, int committed);
 int DPMI_free(dpmi_pm_block_root *root, unsigned int handle);
 dpmi_pm_block *DPMI_realloc(dpmi_pm_block_root *root, unsigned int handle, unsigned int newsize);
-dpmi_pm_block *DPMI_reallocLinear(dpmi_pm_block_root *root, unsigned long handle, unsigned long newsize, int committed);
+dpmi_pm_block *DPMI_reallocLinear(dpmi_pm_block_root *root, unsigned handle, unsigned newsize, int committed);
 dpmi_pm_block *DPMI_mallocShared(dpmi_pm_block_root *root,
         const char *name, unsigned int size, int flags);
 dpmi_pm_block *DPMI_mallocSharedNewNS(dpmi_pm_block_root *root,
@@ -56,13 +56,13 @@ int DPMI_freeShared(dpmi_pm_block_root *root, uint32_t handle);
 int DPMI_freeSharedNS(dpmi_pm_block_root *root, uint32_t handle);
 int DPMI_freeShPartial(dpmi_pm_block_root *root, uint32_t handle);
 void DPMI_freeAll(dpmi_pm_block_root *root, dpmi_pm_block *p);
-int DPMI_MapConventionalMemory(dpmi_pm_block_root *root, unsigned long handle,
-  unsigned long offset, unsigned long low_addr, unsigned long cnt);
+int DPMI_MapConventionalMemory(dpmi_pm_block_root *root, unsigned handle,
+  unsigned offset, unsigned low_addr, unsigned cnt);
 dpmi_pm_block * DPMI_mapHWRam(dpmi_pm_block_root *root,
   dosaddr_t base, unsigned int size);
 int DPMI_unmapHWRam(dpmi_pm_block_root *root, dosaddr_t vbase);
-int DPMI_SetPageAttributes(dpmi_pm_block_root *root, unsigned long handle, int offs, u_short attrs[], int count);
-int DPMI_GetPageAttributes(dpmi_pm_block_root *root, unsigned long handle, int offs, u_short attrs[], int count);
+int DPMI_SetPageAttributes(dpmi_pm_block_root *root, unsigned handle, int offs, u_short attrs[], int count);
+int DPMI_GetPageAttributes(dpmi_pm_block_root *root, unsigned handle, int offs, u_short attrs[], int count);
 int dpmi_lin_mem_rsv(void);
 int dpmi_lin_mem_size(void);
 int dpmi_lin_mem_free(void);
